Move HistoryProviderBase to QuantConnect.Common project
