+++ Engine/QuantConnect.Lean.Engine.csproj
@@ -216,7 +216,6 @@
     <Compile Include="DataFeeds\UpdateData.cs" />
     <Compile Include="DataFeeds\ZipDataCacheProvider.cs" />
     <Compile Include="DataFeeds\ZipEntryNameSubscriptionDataSourceReader.cs" />
-    <Compile Include="HistoricalData\HistoryProviderBase.cs" />
     <Compile Include="HistoricalData\SineHistoryProvider.cs" />
     <Compile Include="Results\BaseResultsHandler.cs" />
     <Compile Include="Results\RegressionResultHandler.cs" />
